single_node.py
from dotenv import load_dotenv
from typing import Annotated, Literal
from langgraph.graph import StateGraph, START, END
from langgraph.graph.message import add_messages
from langchain.chat_models import init_chat_model
from pydantic import BaseModel, Field
from typing_extensions import TypedDict

load_dotenv()

from langchain.chat_models import init_chat_model

llm = init_chat_model(
    "llama3.2:latest", 
    model_provider="ollama",  # Explicitly tell LangChain to use Ollama
    temperature=0
)

class State(TypedDict):
    messages: Annotated[list, add_messages]

graph_builder = StateGraph(State)

def chatbot(state: State):
    return {"messages": [llm.invoke(state["messages"])]}

graph_builder.add_node("chatbot", chatbot)
graph_builder.add_edge(START, "chatbot")
graph_builder.add_edge("chatbot", END)

graph = graph_builder.compile()

while True:
    print("\n\n----------------------------------------")
    user_input = input("Enter a message (enter q to exit): ")
    print("\n\n")
    if user_input == "q":
        break
    state = graph.invoke({"messages": [{"role": "user", "content": user_input}]})
    print(state["messages"][-1].content)