requirements.txt
python-dotenv
langgraph
langchain[anthropic]
ipykernel